+++ app/auth/reset-password/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import ResetPasswordPage from "./_components/reset-password";
+
+export default function Page() {
+  return <ResetPasswordPage />;
+}
